pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "repseek"
description = "Search for plamid replication proteins based on protein structures"
authors = [{ name = "Y.Moriwaki", email = "virgospica93@gmail.com" }]
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    'Operating System :: Microsoft :: Windows',
    'Operating System :: POSIX',
    'Operating System :: Unix',
    'Operating System :: MacOS',
]
requires-python = ">=3.9"
dependencies = ["absl-py", "biopython", "pandas"]
dynamic = ["version"]

[project.optional-dependencies]
dev = ["pytest", "flake8", "mypy", "black"]

[tool]
rye = { dev-dependencies = [
    "ipython>=8.18.1",
    "ipykernel>=6.28.0",
    "pytest>=7.4.4",
    "notebook>=7.0.6",
] }

[tool.setuptools]
package-dir = { "" = "repseek" }

[tool.setuptools_scm]
write_to = "repseek/_version.py"
version_scheme = "python-simplified-semver"
local_scheme = "no-local-version"

[tool.black]
line-length = 100

[tool.flake8]
max-line-length = 100
